OCR/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
pydantic==2.11.3
python-dotenv==1.1.0
requests==2.32.3
Pillow==11.2.1
PyPDF2==3.0.1
PyMuPDF==1.25.5
PyYAML==6.0.2
rapidfuzz==3.13.0
numpy==2.2.5
scikit-image==0.25.2
pytest==8.3.5
# Local OCR fallback
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
